+++ src/main/java/com/rocketapi/app/rest/Models/Elevators.java
@@ -11,7 +11,7 @@ public class Elevators {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     @Column
     private String types;
     @Column
@@ -44,12 +44,12 @@ public class Elevators {
     private int column_id;
 
 
-    public int getId()
+    public long getId()
     {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
 
         this.id = id;
     }
